Huge refactor. Release 0.7.0
Lots of iterations went into this, both locally and upstream with `es_wait`

  - `es_wait` dependency was bumped to 0.5.7

Docker tests have been reworked to create full SSL local instances with captured credentials loaded to a `.env` file.
Docker test containers are no longer made from a project-level image created with small JVM options, as the `-m #GB` option allows easy configuration (heap will automatically be 50% of what is specified after the `-m` flag.  This makes it easy to only have to download and maintain the one image from `docker.elastic.co/elasticsearch/elasticsearch`. In order for tests to work against certain features, a trial license _must_ be enabled. With the short lives of Docker containers, this is not a problem, but the API call must be made. I am currently doing it at client object creation time in my pytest `conftest.py` file. But should that prove cumbersome, I will probably move it to the `docker_test/create.sh` script and just have a line of `curl` finish it at the end. For right now, though, since repository creation is a bit easier in `conftest.py` since I have the planned name of the repository, etc., I'm handling it there. If I really want to centralize it in `docker_test` scripts, I can. Just a few more curl calls. Teardown of those already deletes the bind mount volume for the repository anyway, so why not. Consider it a to-do.

A huge refactor of TestBed teardown was enacted, making it all local in the end. No more "tracker" class.
A minor refactor of TestPlan was required to ensure that ILM was being configured properly (more on ILM below).

A refactor of the ILM advance of phases and steps was required. It was too inconsistent, especially with how fast I want the setup/creation process to be, whether on Cloud or local to Docker. It is astonishing how many tiny steps are involved with the progression from new -> {phase: hot}, leave alone to {phase: hot, action: complete, name: complete}. Lots of further steps to get to {phase: cold} (then complete), or {phase: frozen} (then complete). The `es_wait` module was updated to address these with code started in `es_testbed`. A lot of iteration there to make that work as expected. Most of the difficulties I encountered were ILM related, in fact. As a result, the following note:

  - **NOTE:** It was discovered that at least in 8.13.2, local Docker instances cannot and/or do not work with ILM promotion from hot/warm directly to frozen. The cluster just locks up with the frozen index "created" but in a perpetually red state, and the original index is also just there still, unable to go anywhere.  For this reason, the ILM promote to frozen tests are skipped if you are testing against localhost:9200. I have not had time to chase down the "why" of this. I tried increasing the memory/heap, to no avail. It may require 2 nodes. The single docker instance still had every node role, so it wasn't a function of node roles (manual promotion to frozen still works, just not ILM promotion to frozen). It may be a weird bug I need to report. Will figure it out at some point.

With regards to logging, logging is now at the module level, period. Too many class inheritance issues (while using one logger per class, e.g. self.logger, and passing it to child classes) were making it harder to see which code was being executed by the parent class code or a child classes extended method.

To shorten module name paths so log lines are a bit shorter, class files were re-named:

  - src/es_testbed/mgrs/componentmgr.py -> src/es_testbed/mgrs/component.py
  - src/es_testbed/mgrs/data_streammgr.py -> src/es_testbed/mgrs/data_stream.py
  - src/es_testbed/mgrs/entitymgr.py -> src/es_testbed/mgrs/entity.py
  - src/es_testbed/mgrs/ilmmgr.py -> src/es_testbed/mgrs/ilm.py
  - src/es_testbed/mgrs/indexmgr.py -> src/es_testbed/mgrs/index.py
  - src/es_testbed/mgrs/snapshotmgr.py -> src/es_testbed/mgrs/snapshot.py
  - src/es_testbed/mgrs/templatemgr.py -> src/es_testbed/mgrs/template.py

A ``.coveragerc`` file was added.

Testing has only been mildly modified. Not all code has coverage yet, but that will come with time.
Also of note, the tests are not automatically spinning up and tearing down Docker images, _yet_. That step will come later.