Release prep for 0.8.4
Most of this change is in the form of using es-docker-test-scripts version 1.2.0

The 1.2.0 version of the `docker_test` scripts proved necessary to run two docker containers
as part of a single cluster, with one of them being a `data_frozen` node.

This is accomplished by running:

```
docker_test/create.sh 8.15.0 frozen_node
```

where `frozen_node` is the scenario defined in the docker_test scripts for deploying
a 2 node cluster with one `data_frozen` node.

With that out of the way, the other changes were to the searchable_test preset scenario.

The import of this scenario seemed to require class attributes for the desired settings to
be properly exposed, so `scenarios.py` was changed to include a class `Scenarios`, with every
possible scenario as a class `@property`/attribute.

Testing showed that these values properly propagated.

Now, with both changes in place, the prior failure to get cold and frozen ILM tests to work is a thing of the past.